moving components to subdir
